backend/requirement.txt
fastapi==0.111.0
uvicorn==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest==8.2.0